requirements.txt
requests
pywebview
websocket-client
